move create movedoc api and ducks to movedoc dir
