+++ src/scenes/Office/DocumentViewer/api.js
@@ -8,3 +8,26 @@ export async function IndexMoveDocuments(moveId) {
   checkResponse(response, 'failed to get move documents due to server error');
   return response.body;
 }
+
+export async function CreateMoveDocument(
+  moveId,
+  uploadIds,
+  title,
+  moveDocumentType,
+  status,
+  notes,
+) {
+  const client = await getClient();
+  const response = await client.apis.moves.createMoveDocument({
+    moveId,
+    createMoveDocumentPayload: {
+      upload_ids: uploadIds,
+      title: title,
+      move_document_type: moveDocumentType,
+      status: status,
+      notes: notes,
+    },
+  });
+  checkResponse(response, 'failed to create move document due to server error');
+  return response.body;
+}
